[ShellScript] Fix substring replacement cases.
Test cases from okdana:

Parameter-expansion operators should be treated as such only where
appropriate. They should NOT be treated as unqualified operators anywhere they
appear in an expansion construct.

: ${foo//%/} # '%' is not an operator here
: ${foo//#/} # '#' is not an operator here
: ${foo//!/} # '!' is not an operator here
: ${foo//:/} # ':' is not an operator here
: ${foo//@/} # '@' is not an operator here

When using the replace operators ('/' and '//'), escaped and un-escaped
slashes in the needle and replacement strings should be handled correctly.
Slashes need be escaped only in the needle; escaping slashes in the
replacement is optional.

: ${foo//a\/b/c/d}
       ^^    ^     These slashes are operators
           ^   ^   These are not