[C] Fix issue with macro crossing block
Fix Issue #201 where a macro crossing a block { } would break
highlighting.
I simply removed the context push which only served the purpose of
highlighting #endif only if it followed a #if: so this code could
highlight some invalid code, but at least it does not break the
highlight for valid code.