[Python] Add scope for debugging
Uncomment this to spot characters that fall through all else
(such as stray braces or assignments in expressions).

In order to not remain forward compatible, this match is not active.
Also prevents matching something as invalid that actually is valid but
the syntax definition does not match yet
(such as keywords in class inheritance).