[Python] Consider ; statement termination
All except for one use of the line_continuation include were statements
that would otherwise end at the end of the line, before a comment or
before a semicolon. (Eventually, that should be eliminated as well once
assignments and target-lists are specifically matched.)
Thus, combine these into a single "line_continuation_or_pop" context and
include that at several places to fix them all.

A test was only added for imports.