Fix scope mismatch between Ruby syntax and Haml
I found this while investigating all the scopes in the repo where order mattered and this was the only one. Turns out it is a mistake and this should match the one in the Ruby syntax file.

So fun fact: in all the default syntaxes order of the scope doesn't matter when checking equality. Not sure if this is true for all prefixes yet though.