[XML] Permit unescaped > in content
Greater-than is actually valid in both text content and attribute values
(see CharData and AttValue productions). Change should-be-entity to reflect
this.

The exception is that the sequence ]]> must not appear in text content (but
it's OK in an attribute value, for some reason). Add missing-entity detection
for that inside text content.