[Python] Introduce qualified-name-until-leaf ctx
... and use it to match function calls less explicitly and more
magically. Usually you would think this is bad, but it's not because
it's not really less explicit. It just avoids repetition when scoping
the last (non-special) part of a qualified name with a name other than
"meta.generic-name".

This change triggers the bug described in
https://github.com/SublimeTextIssues/Core/issues/1190, where the
most-inner context in a multi-push operation *must* match at least one
character before it can be popped.

Due this, the tests are currently *not* passing.