Add support for Objective-C format specifier
Objective-C and Objective-C++ support the standard format specifiers
in format strings as well as an additional format code of `%@` which
designates the string representation of an object as determined by one
of its description methods.

This change adds in detection for this format specifier in object
strings and includes two new syntax tests to ensure that the specifier
is allowed in an NXConstantString but not in a standard C string
constant.

In my tests, this change to the Objective-C syntax is also applied to
files of type Objective-C++ with no further changes.