Fix highlighting of __cplusplus macros without breaking correct pop off from external C.
The deleted block from the previous commit was added
by commit 0e7352481ca760c8d8939f2962dcd3dde94333b2
to fix an issue with correctly popping off from an external C block.

Instead of deleting the block adding `preprocessor-global` and `global` to it also
fixes the highlighting issue, but does not break the previous fix.

Added this to the testfile, too.