[HTML] indentation rule improvements
allow text (i.e. excluding angle brackets) preceding the close tag for unindent rule.
indent the next line if there is an unclosed open tag on the line.
use `\k` rather than `\g` to reference the previously captured named group because (in Oniguruma), the latter recurses the expression and not the match.