[XML] Highlight some chosen characters in strings (DTD)
Some operators and characters as well as embedded quoted strings are highlighted as constant.character.escape to make ENTITY values more readable. It is not yet intended to support full syntax highlighting in ENTITY values as they can contain too divergent data which causes highlighting to fail sometimes. Furthermore a string is a string and should be realized as such.