[XML] Reorder some rules to apply a kind of top-down-order.
It seems a good practice to sort lexer rules a little bit. Scopes with more general/reusable functionality go downwards while highlevel functions keep on top beginning with main. On the other hand why should strings not be defined next to each other? They don't need to be spread in the file.