[CSS] Use \x{} for all custom_element_chars
Some regex engines such as Oniguruma distinguish between `\x..` and
`\x{..}` in that the former matches bytes whereas the latter matches
Unicode code points, see here:

https://github.com/kkos/oniguruma/blob/master/doc/RE#L26

The CSS syntax matches e.g. the U+00B7 character. So for that to work in
Oniguruma, it needs to be written as `\x{B7}`. For readability, this
change uses `\x{00B7}` which is closer to how Unicode characters are
commonly denoted.