[C++] [Objective-C] Add standard library include completions
Had to adjust the way the ObjC syntax scopes "#import ..." for this. It
used to assign a meta scope of `meta.preprocessor.include`, but it now
assigns a meta scope of `meta.preprocessor.import`. The "#include ..."
statements still receive `meta.preprocessor.include` for their meta scope.
This is done so that the completions only trigger for an #include and not
for an #import.