[YAML] Fix timestamp highlighting with space separated timezone
The example.yaml of the PyYAML package and the example on http://yaml.org/type/timestamp.html both show a timestamp with space separated timezone:

  2001-12-14 21:59:43.10 -5

It is not matched by the regexp displayed on the same page which is used by YAML.sublime-syntax to do highlighting. But as a canonical timestamp like

  2001-12-15T02:59:43.1 Z

is considered valid, there is no reason to not match the first one, too.