[Git] Commit: Add support for squashed commit messages.
Committing a squashed or fixed up commit, the message text looks like in the added `syntax_test_git_squash` file.

The message starts with a comment followed by a empty line. Therefore we need to change the way, the commit subject is detected.
Therefore the first none empty none comment line is matched as commit subject.