[Git] Commit: Less strict empty 2nd line checks
The commit syntax file may be included by `git log.sublime-syntax` in the future, which displays the commit message indented by 4 chars. In order to prevent unwanted invalid highlighted empty second line because of whitespace at the beginning of the line, the 2nd line is expected empty, if it contains whitespace only.