[Git] Rebase: Prepare for common reference scopes
References such as issues, username, email are used in several contexts.
Therefore they may be outsourced to a common hidden syntax file and to
be imported wherever needed. This commit prepares for such an operation,
by reusing the rules from gitcommit.sublime-syntax.

In the future the `references` may be moved to a `Git Common.sublime-syntax`.