[Python] Highlight `_` as language variable
`_` in itself does not have a special meaning, but it is a
conventionally used identifier much like `self` and `cls` for when you
don't want to store a value into a variable but are required to do so
syntactically.

This is usually the case for unpacking/pattern matching tuples or lists,
or in for loops.

Additionally, `_` has a special meaning in an interactive session where
it refers to the value returned by the last expression executed.