[CSS] Add a hint to each completion item
Most auto-completion packages provide basic information about their
suggestions as hints right next to each item, which distinguishes
auto-completion items very clearly from word-completions.

CSS disables word-completions, but it is not obvious to the end user.

Therefore this commit adds a hint to each returned completion item to
allow the end user to clearly see it is a valid value from a database
but not something ST might have caught from anywhere in the text.