[Git Files] Config: Increase performance by 13%
This commit replaces the 'expect-key' rule which was responsible to pop off from a value scope by a normal 'line-end-pop'. The implementation of the trailing backslash as 'line-continuation' was taken from the default Python syntax.

To make the rule-names match a naming scheme after that change, some further renames were performed:

- expect-newline		=> expect-line-end
- illegal-newline		=> illegal-line-end
- illegal-newline-pop	=> illegal-line-end-pop

As a side-effect the key-value-pair rules get a bit simpler, too.

The overall performance was benchmarked with a .gitconfig file with 100.000 lines of code.