[PHP] Add highlight support for phpdoc tag source
The "source" tag can be used to communicate the implementation of
Structural Elements by presenting their source code, or more typically -
portions of it.

If specified, the starting line MUST be a positive integer. Counting
starts at the line where the Structural Element's body started (i.e. the
line with the opening brace). This line number MAY optionally be
followed by another positive integer, specifying the number of lines to
present from the starting line onwards. If omitted, the Structural
Element's source is presented from the starting line, to its end (i.e.
the line with the closing brace).

If no starting line is specified, the Structural Element's source is
presented in its entirety. If no starting line is specified, the
Structural Element's source is presented in its entirety.

Ref: https://docs.phpdoc.org/references/phpdoc/tags/source.html