[PHP] Add highlight support for phpdoc tag method
The "method" tag is used in situation where a class contains the
`__call()` magic method and defines some definite uses.

An example of this is a child class whose parent has a `__call()` to
have dynamic getters or setters for predefined properties. The child
knows which getters and setters need to be present but relies on the
parent class to use the `__call()` method to provide it. In this
situation, the child class would have a "method" tag for each magic
setter or getter method.

The "method" tag allows the author to communicate the type of the
arguments and return value by including those types in the signature.
When the intended method does not have a return value then the return
type MAY be omitted; in which case `void` is implied. The "method" tags
MUST NOT be used in a PHPDoc that is not associated with a class or
interface.

Ref: https://docs.phpdoc.org/references/phpdoc/tags/method.html