[PHP] Add highlight support for phpdoc tag filesource
The "filesource" tag tells phpDocumentor to include the current file in
the parsing output. As this only applies to the source code of the
entire file MUST this tag be used in the file-level PHPDoc. Any other
location will be ignored. When this tag is included will phpDocumentor
compress the file contents and encode them using Base64 so that it can
be handled by the transformer. Any template that is able to show the
source code can then read the source sub-element in the associated file
element in the Abstract Syntax Tree.

Ref: https://docs.phpdoc.org/references/phpdoc/tags/filesource.html