[PHP] Add highlight support for phpdoc tag property
The "property" tag is used in the situation where a class contains the
`__get()` and `__set()` magic methods and allows for specific names.

An example of this is a child class whose parent has a `__get()`. The
child knows which properties need to be present but relies on the parent
class to use the `__get()` method to provide it. In this situation, the
child class would have a "property" tag for each magic property. The
"property" tags MUST NOT be used in a PHPDoc that is not associated with
a class or interface.

Ref: https://docs.phpdoc.org/references/phpdoc/tags/property.html