[PHP] Add highlight support for phpdoc tag ignore
The "ignore" tag tells phpDocumentor that the Structural Elements
associated with the tag are not to be processed. An example of use might
be to prevent duplicate documenting of conditional constants. It is
RECOMMENDED (but not required) to provide an additional description
stating why the associated element is to be ignored.

Ref: https://docs.phpdoc.org/references/phpdoc/tags/ignore.html