[repo] Add .editorconfig to repo root (http://editorconfig.org)
To use this an editor needs to support it. For SublimeText there
is a package called EditorConfig (I'd argue this should become
part of ST Core in future). For other editors, see editorconfig.org
for a list of editors that have built-in support and ones that have
plugins.

This commit only provides a first draft as starting point based on
what a first time contributor thinks the settings should be. This
file was prompted after the user-specific setting
"trim_trailing_white_space_on_save" caused a syntax test to break
(it had specific trailing whitespace for a test).