[Groovy] Fix invalid highlighting after embedded code in string
Fixes issue #1451

This commit fixes an issue with the closing `"` of a  string being matched as start of a new string, which caused the region after it to be scoped as invalid.

The reason for this issue was the end of the embedded code block `${ ... }` not being recognized due to the used expression inside, which expected the code block to end at the end of line only.