[Python] Fix invalid highlighting of closing brackets in generators
Fixes #1505

In generator expressions after `for` a `target-list` section is used to scope the temporary variables. This section highlights names only and therefore fails in case of nested groups. The result is described in #1505.

This commit enables nested groups. Instead of using `groups` another `target-list` is included for nesting with the assumption of those nested groups not supporting all kinds of expressions anyway.

Furthermore the opening and closing parentheses are scoped as `punctuation.section` to be addressable by color schemes.