[ShellScript] Fix vardef-name context not popping
For example, in
```
printFunction "$variableString1" "$(declare -p variableArray)"
```
the token `variableArray` is scoped with the vardef-name context, and it should
pop when encountering the closing parenthesis. We solve this by looking ahead
for a metacharacter.

In general more gratuitous popping in the vardef-* contexts is applied in order
to not get stuck in an unwanted context.