[XML] Fix DTD comment prototypes
## Issue

The `prototype` section in the DTD.sublime-syntax file is ignored,
if the syntax is included/embedded into another (XML) syntax.

As this section contains the `comment` rules, they are not applied
if DTD is embedded somewhere else.

Up to this commit these comments were scoped by the encapsulating XML,
which works seamlessly but may cause issues, if DTD is embedded into
a syntax, which does not support this kind of comments or does not
provide the proper `with_prototype` statement.

## Solution

In order to fix that issue, the `prototype` is explicitly included
into DTD's `main` section. As a result the `with_prototype` statement
in the `doctype` section of the XML syntax can be removed.