[HTML] More precise html-end-tag scoping
Whitespace is no punctuation and should therefore not scoped as such.

This commit changes the `tag-end-maybe-self-closing` to no longer
include the preceding space into the `punctuation.tag.end.html`.

The reason for it to be included here is not obvious anyway as none,
one or any number of spaces would cause the `/?>` to be detected as
`end of tag` anyway.