[JavaScript] Fix function declarations within objects break highlighting
Fixes #1593

### Issue

According to the minimal test case, if `.a()` is not followed by an
opening `{` (a real `function-block`), the closing `}` of the owning
object is matched as closing punctuation of the none present function
block.

As a result the context of the owning object (`meta.object-literal`)
is never popped off anymore.

### Reason

The reason for that issue is located within the `function-block` context,
which consumes the next available `}` after a function declaration
without respect to a prior `{` to be required.

### Solution

The code structure and the use of the `function-block` context is
adapted to `function-declaration-parameters`.

It now contains the rule to set the `meta.block` context only.
An opening `{` is required before looking for a closing `}` for it.
By removing the `else-pop`, some look aheads are removed from

  - `function-declaration-expect-body`
  - `arrow-function-expect-body`