[XML] Refactor syntax test test case comments
This commit introduces `##` as punctuation for syntax test cases
and adds a prototype context to the XML.sublime-syntax file which
can be commented in or out in order to highlight these comments as
such or ignore them.

The main goal is to introduce a comment style which is independent
from valid XML style comments as those might break otherwise valid
xml code when added in certain places for testing.

Example:

  XML comments are not valid within xml tags. In order to check,
  whether xml tags work perfectly if spanning multiple lines,
  test cases must be added within tags, which would then break them.

The `##` is choosen as it is used by the HTML.sublime-syntax as well.

Note:
  The `protoype` context should be commented out before committing!