[XML] Fix broken highlighting after incomplete tags/attributes
The main goal of this commit is to improve user experience while
writing XML documents by not breaking the highlighting of content
which follows incomplete tags or attributes.

It adds an `tag-end-missing-pop` context, which is used to pop off
from a tag even though it is not yet closed, as soon as the next
opening tag punctuation `<` is detected.

As a hint the preceding whitespace before `<` is highlighted as
`invalid.illegal` without breaking the rest of the document.

A couple of test cases are added illustrating the new behavior.