[HTML] Fix internal subsets in DOCTYPE break highlighting
Fixes #1565

If the DOCTYPE tag contains an internal subset to extend the scope
of available xml entities, syntax highlighting breaks because the
closing `>` of subsequent tags is matched as end of the DOCTYPE.

The current html5 specifies the use of

  <!DOCTYPE html>

without any other tags inside, but older standards may require more
information to be provided, such as

  <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.1//EN"
                        "http://www.w3.org/TR/xhtml11/DTD/xhtml11.dtd">

see: https://html.com/tags/doctype/

Even though internal subsets are not proposed as valid for html,
they at least should not break highlighting as the HTML syntax gets
included in more and more special purpose syntaxes, which may include
more sophisticated DOCTYPE tags.

This commit adds a `doctype` context to properly highlight all DOCTYPE
tags mentioned in https://html.com/tags/doctype/ to satisfy the older
html4 or xhtml specs, too.

The content of the internal-subset is not handled in any way at the
moment.

FUTURE

In the long term `comment`, `preprocessor` and `doctype` contexts
should be implemented in detail in the XML.sublime-syntax and included
here as none of the three is dedicated HTML.