[HTML] Scope CDATA as declaration
## Issue

The <![CDATA[..]]> tag is completely scoped as string which makes
their text content hard to distinguish from the enclosing tag elements.

## Solution

Inspired by the previous commit, the CDATA tag is handled as declaration
in its own context. The punctuation and CDATA keyword are adopted to the
DOCTYPE declaration tag.

## Remarks

As the html syntax is more and more included into special purpose
syntax definitions, all other <! .. > tags are no longer handled
at all in order reduce the risk of interfering with the owning
syntax definition.