[Perl] Allow any kind of whitespace around accessors
The accessor `::` may be surrounded by spaces and tabs. Not sure whether
someone makes use of tabs in such situations, but it should not fail if
one decides to do so.

HEREDOC indention still accepts normal spaces only at the moment as the
amount of spaces in front of the starting and terminating keyword must be
the same which might be tricky to handle if mixed indention is used.