[C++/ObjC] Fix: Parens inside template arguments
For things like std::function<A<B<C>>()> f(g), the grammar got confused and
thought this was a templated function call (the function name being "function"
and the start of the opening parenthesis for the function call being the
parenthesis after the "f" identifier). However, this is a new variable called
"f" of type "std::function<A<B<C>>()>" initialized with the variable "g".

To address this, the generic_lookahead variable had to be changed to account
for possible opening and closing parentheses inside of its template argument.

The result is a somewhat insane generic_lookahead, but it makes the syntax more
correct.