[Diff] highlight diff section hints
The toc-list of a changed hunk in a patch file can contain hints about the context of a change - maybe a function name.

As this hint was not yet respected in the diff syntax, the hint was highlighted as normal text. If a color scheme defines a background color not the whole line was highlighted.

Hence this commit adds support to properly highlight section hints.