[#1769] Fix for Clojure code block
There was an issue in that the rendering of the Clojure code block in
a Markdown file would not be rendered properly. This is documented in
Issue #1769 - with screen grabs. This addition to the syntax file
corrects that issue by adding the section for the Clojure code block in
the Markdown file.