[HTML] Improve compatibility with template languages
This commit introduces changes to the custom-/other-tags parsing in order to enable template languages like ASP/PHP/JINJA to easily inject their template tags into the middle of a tag-name.

Another benefit of this approach is the ability to highlight single invalid characters within the tag-name instead of scoping everything as invalid.illegal.