[HTML] Improve compatibility of attributes with template languages
This commit applies a suggestion from @Thom1729 about improving coexistence of html with other syntaxes which are based on it.

Problem Description

A general problem is that languages like ASP or PHP or even JINJA might inject special tags into tag-names, attribute-names or attribute-values, like

   <my<? $foo ?>tag a<? $ttr ?>ibute=va<? $lu ?>e />
    ^^^^^^^^^^^^^^^ tag
                    ^^^^^^^^^^^^^^^^ attribute
                                     ^^^^^^^^^^^^ value

This is only possible if each word is matched in its own context instead of using one monolithic pattern as the start-token of the injected code-snippet is unknown.

Solution

This commit therefore applies the strategy which is already used in the `tag-generic-name` context to unquoted-attribute-values as well.

Another benefit of this approach is the ability to highlight single invalid characters within the attribute-name or value instead of scoping everything as invalid.illegal.

The negative lookahead in the `tag-generic-attribute-name` is replaced by a positive one in order to improve parsing performance. Looks like ST's sregex engine can handle the latter one better.

A quick check with a modified version of Celestre.sublime-color-scheme was used to make sure attribute names/values are compatible with the hashed highlighting of ST.