[JavaScript] Move JSON out of the JavaScript package
As suggested in issue #1771 this commit moves the JSON syntax into a
dedicated package in order to be able to disable either JavaScript or
JSON independently.

This commit does not yet create the suggested compatibility
redirection within the JavaScript package.

The decision about it is left to the core devs right now

1. as it would cause a duplicated syntax file
2. a quick search within packagecontrol.io did not reveal any packages
   depending on the builtin JavaScript/JSON.sublime-syntax
3. the location of the sublime-syntax file should not be too important
   if it is imported via `scope:source.json` or its name
   `JSON.sublime-syntax` only. Conflicts would be caused only if it was
   imported via `Packages/JavaScript/JSON.sublime-syntax` which is easy
   to fix.