[C#] Fix lambda operator not matched in function arguments
Before this commit the lambda operator `=>` was not matched within function call arguments like `func(arg => anonymous)`.

To fix the issue the negative lookahead `(?!=)` in the `arguments` context is extended to restrict the matching `=` as assignment. A positive lookahead with a negated character class is used as ST's regex engine can handle that much faster than negative lookaheads.