[ShellScript] Fix case context stacking
This commit fixes a design issue of the case context, which could probably cause the parser to hit a sanity limit if a file with too many case clauses was loaded.

Instead of chaining each case clause via `set` keyword until `esac` is matched, a common `case-body` context is introduced from which each pattern/command-list pair is pushed. This results in at most two contexts to be pushed to the scope stack no matter how many clauses the case statement contains.