[ShellScript] Highlight stray `esac` as keyword like stray `fi`.
If the leading `case` keyword is missing or incomplete, the finalizing `esac` should be scoped as keyword as well.

This commit does not scope it as illegal as it is not yet implemented for `fi` as well. The syntax does not handle many illegals at all, so this commit does not intend to change that.