[Python] Fix Oniguruma incompatibility with regex for path
In 0550509bcd5, `path` was modified to allow spaces around `.`. Later,
`simple_expression` started using `path` but within a `(?x:...)`, which
means this:

    \b[[:alpha:]_][[:alnum:]_]*\b *\. *

is equivalent to:

    \b[[:alpha:]_][[:alnum:]_]*\b*\.*

While sregex doesn't complain about that, `\b*` is invalid in Oniguruma.
Note that even with sregex, the meaning of the regex changes.

By using a character class for the space, the pattern works the same way
whether it's within a `(?x:...)` or not.