[Perl] Fix predefined sub-routines
This commit moves the keywords `AUTOLOAD` etc. out of the functions context to the `sub` context as these names describe special sub routines a user can define as callback to be called by perl in certain situations.

This step is not crucial to fix issue #1902, but is found worth doing while investigating for further steps to fix the main issue.