[Perl] Fix user defined sub-routines
The issue #1902 is triggered by recognizing a member value of an object as member function, while user defined functions were assumed to be called with the same syntax as builtins.

But while arguments of builtin functions like `print`, `say`, etc. don't need to be surrounded by parentheses, the ones of user defined sub routines must be.

Examples:

  Valid builtin function call:

    print <Hello World>

  Valid user defined sub routine call:

    sub myprint(@arg) { print @arg; }

    myprint(<Hello World>);

This commit makes sure to match a token as user defined function call only if the identifier is followed by parentheses and adds `meta.function-call` scopes as the parentheses make it easy to find the arguments boundaries.

It does not add `meta.function-call` scopes to builtin function calls. Due to some limitations with regards of pushing into contexts which are required to handle HEREDOCS correctly it is not possible to find the end of an argument list reliably.