[Perl] Distinguish function/variable object members
This commit is the final step to fix issue #1902.

It changes the `->` token from `keyword.operator` to `keyword.accessor` and introduces distinguishing object variable and function members, because the most recent commit caused a token like `->value` to not be scoped at all anymore.

The Perl syntax now scopes `->value` as `variable.other.member`  and `->func()` as `variable.function.member` as it was introduced for class members a couple of commits before.