[Perl] Fix member function vs. member variable matching (#1911)
* [Perl] Distinguish function/variable class members

Issue #1902 was introduced by commit 8726418 and is triggered by not correctly distinguishing function and variable members of classes or objects.

This commit therefore starts with member matching of classes as a first step to fix issue #1902.

* [Perl] Fix predefined sub-routines

This commit moves the keywords `AUTOLOAD` etc. out of the functions context to the `sub` context as these names describe special sub routines a user can define as callback to be called by perl in certain situations.

This step is not crucial to fix issue #1902, but is found worth doing while investigating for further steps to fix the main issue.

* [Perl] Fix user defined sub-routines

The issue #1902 is triggered by recognizing a member value of an object as member function, while user defined functions were assumed to be called with the same syntax as builtins.

But while arguments of builtin functions like `print`, `say`, etc. don't need to be surrounded by parentheses, the ones of user defined sub routines must be.

Examples:

  Valid builtin function call:

    print <Hello World>

  Valid user defined sub routine call:

    sub myprint(@arg) { print @arg; }

    myprint(<Hello World>);

This commit makes sure to match a token as user defined function call only if the identifier is followed by parentheses and adds `meta.function-call` scopes as the parentheses make it easy to find the arguments boundaries.

It does not add `meta.function-call` scopes to builtin function calls. Due to some limitations with regards of pushing into contexts which are required to handle HEREDOCS correctly it is not possible to find the end of an argument list reliably.

* [Perl] Distinguish function/variable object members

This commit is the final step to fix issue #1902.

It changes the `->` token from `keyword.operator` to `keyword.accessor` and introduces distinguishing object variable and function members, because the most recent commit caused a token like `->value` to not be scoped at all anymore.

The Perl syntax now scopes `->value` as `variable.other.member`  and `->func()` as `variable.function.member` as it was introduced for class members a couple of commits before.