Only index variables and members longer than 4 characters
Indexing common words is pretty useless. It also slows down sublime a
lot when hovering over a common variable name.

I just added a .indexed to the scopes. Not sure if this will fly. It
might be better to add a meta.indexed scope to all the variable.other
and entity.name.members that should be indexed.