[Perl] Improve escaping from embedded regexp
Nearly any character can be used as delimiter in quoted-like
expressions.

All the following expressions are equal:

   s/<pattern>/<repl>/<flags>
   s|<pattern>|<repl>|<flags>
   s#<pattern>#<repl>#<flags>
   s@<pattern>@<repl>@<flags>

The usage depends on the characters being used in the <pattern> or
<repl>. If `/` is used in must not be part of them. Perl however
accepts escapes like `\/`.

Theoretically a <pattern> can end with an arbitrary number of `\`,
which makes creating a robust `escape` pattern complicated or even
impossible.

Before this commit the escape patterns check for the absence of a
single `\` before the delimiter candidate `/` only.
The check is extended to `\` and `\\\` in order to allow windows style
path names ending with an escaped backslash.

This is not a perfect solution, but should help in some rare edge cases.