[Perl] Fix identifier break
An identifier can be followed by a double-colon. Hence using the
`{{break}}` is not appropriate and needs to be fixed.

This is a theoretical issue which is not triggered at any point of the
current implementation. This commit therefore has no functional effect
to the current syntax definition, but may have due to future changes.

Note: This change is covered by existing test cases.