[Perl] Add sprintf format placeholders
This commit

1. adds the `literal-common` and `interpolated-common` contexts to
   group common content for interpolated and raw strings.
2. adds rules to scope C-style FORMAT placeholders as being used by
   printf/sprintf, etc..

   See: https://perldoc.perl.org/functions/sprintf.html

The pattern to match the FORMAT is designed quite restrictive in order
to prevent interference with variable interpolation. Perl uses `%` for
some types of variables and the FORMAT is to be supported in any kind
of quoted string as it might be the format string is build dynamically.

Example:

  $format = "%s: %d";
  sprintf $format, $var1, $integer;

Note:

  Perl supports variable interpolation even in FORMAT patterns.
  Something like %0${varname}X is valid. This commit does not support
  such constructs as it is nearly impossible to distinguish between
  FORMAT patterns and hashs - both start with `%`.