[Perl] Fix HEREDOC function call arguments
Issue:

If a HEREDOC is used as argument in a function call, the arguments
context never pops off. Hence `meta.function-call.arguments` keeps on
stack forever.

Example:

   function_name(<<"   HEREDOC");
   This is the HEREDOCs value which is passed to the function
   HEREDOC

Solution:

This commit removes any `meta.function-call` scopes and all the related
contexts in favor of properly highlighting HEREDOC like arguments in
functions.

As the function call can be nested and followed by arbitrary perl
expressions up to the end of line, no better solution was found with
the existing features of ST's lexer.