[Perl] Fix qualified functions and variables
(I) Function calls without parentheses

Perl Documentation says: Encapsulating function arguments into
parentheses is optional, if the expression is clearly identified as
function-call.

This statement applies to calls of sub routines which are already known
to the interpreter by defining them at the beginning of the script or
by importing them via `use ...` statement from other modules.

This commit adds some heuristics to identify such kinds of function
calls even though we don't really know about the validity of the
identifier.

A function is clearly identified if the `identifier` is followed by
  - comment
  - end of line
  - end of expression (closing bracket)
  - end of statement (semi-colon)
  - HEREDOC
  - quoted string
  - regular expression
  - variable
  - word but no operator

(II) Each variable or function identifier can contain a qualified path.

Example:

  $NS1::NS2::variable
   NS1::NS2::function

The `main` namespace is shortaned by leading `::`

  $::variable
  ::function

This commit applies correct scopes for all qualified identifiers.

(III) Qualifiers or identifiers can consist of only capital letters.

In order to distinguish them from global constants this commit adds
the following rules:

1) Prefer scoping the file handles like STDERR, STDOUT, etc. as
   constants even though they are used as function or look like a class
   or namespace. Even though these tokens don't have special meaning
   to the Perl interpreter they are considered for certain use by the
   standard library.
2) If a user defined constant looks like a function it is scoped as
   such. In other words tokens with capital letters only are scoped as
   constant only, if they don't look like a namespace, class, object or
   function.

(IV) Object member functions

According to https://perldoc.perl.org/perlobj.html an object in Perl
is nothing else than a normal data structure (array, hash, ...) which
is bound to a class.

The `->` accessor always indicates a member function call, if followed
by an identifier. Attributes are accessed via getter/setter methods in
modern Perl code.

Example:

   $obj->method
   $obj->getter
   $obj->setter <value>

A subroutine always returns a reference to an object. Therefore the
following is invalid:

   $obj->method{key}

Another `->` accessor is needed to access the item of the returned hash by `method`:

  $obj->method->{key}

Otherwise the `->` could also be used to directly access data items in
the underlying hash of an object (which is discouraged).

Example:

   $obj->{attribute}

Accessing members of a nested object look like:

   $obj->{attribute}->method

Note:

  Object members like `->new()` in an expression like `Class->new()`
  are not yet part of meta.path as they must not be part of a string
  interpolation.

The `->` accessor scope is renamed to `punctuation.accessor.arrow` to
comply with scope naming guidelines and give it the same color as the
`::` operator which is used to access classes.