[Erlang] Fix anonymous function definitions
This commit applies suggestions from the review process.

Including anonymous function definitions into goto definition features
makes not much sense on the one hand and brings too much noise on the
other. The anonymous function maybe assigned to a variable, but this is
a dynamic reference to the function then, which can be redefined at any
point.

Hence scoping the l-value of an anonymous function definition as
`entity.name.function` is removed. It's scoped as ordinary
`variable.other` instead.