[Erlang] Fix record field scopes
As the local creation/copy of records is no longer scoped as definition
the logical result is to handle the field access within the creation
expression as normal field-write-access instead of a field definition
as well.

To keep the field definition within preprocessor expressions intact,
the `record-body` is now assigned to record definitions only. A copy
and modified `record-fields` context is used within record creation
expressions.

Example:

    Expr#rec{field=1}
             ^^^^^ variable.other.field

Note:

  We might probably think about more general scopes for fields and
  records?

  Maybe:

   `variable.other.field` -> `variable.other.member`
   `variable.other.record` -> `variable.other.struct` or `variable.qualifier`??